Uncommented .env configuration in .gitignore
Previously, the .env configuration was excluded from version control because it was listed in the .gitignore file. In this commit, to include the .env configuration in version control, a comment has been added. This change is done intentionally as it allows allow us to share and track environment configuration changes related to NVIDIA's L4T platform and TensorFlow